+++ database/migrations/2022_03_26_115518_create_orders_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->foreignId('customer_id')->constrained('customers')->onUpdate('cascade')
             ->onDelete('cascade');
             $table->string('payment_type');
-            $table->integer('payment_amount');
-            $table->integer('profit');
+            $table->decimal('payment_amount');
+            $table->decimal('profit');
             $table->string('delivery_company')->nullable();
             $table->string('tracking_number')->nullable();
             $table->string('delivery_name')->nullable();
