+++ database/migrations/2022_03_26_120547_create_order_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('order_products', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('quantity');
+            $table->integer('quantity');
             $table->foreignId('order_id')->nullable()->constrained('orders');
             $table->foreignId('product_id')->constrained('products')->onUpdate('cascade')
             ->onDelete('cascade');
