+++ database/migrations/2022_06_27_050926_create_receiving_items_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             ->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products')->onUpdate('cascade')
             ->onDelete('cascade');
-            $table->tinyInteger('quantity');
+            $table->integer('quantity');
             $table->timestamps();
         });
     }
